Create new_number_game.py
Recreated the game with some of the old code. Made it much simpler using a while true loop and also figured out the high score issue I was having. All the new game files are in the new_number_game.py and I will be deleting the old file to prevent confusion.